chore: Remove modelcif dependency and related code
Deleted the modelcif package from pyproject.toml and removed the
to_modelcif function along with all imports and references to modelcif
in deepfold/common/protein.py. Also committed the removal of poetry.lock
file and unignored poetry.lock in .gitignore to ensure lockfile is not
tracked. This cleans up unused code and dependencies.